优化了贪心算法和控制算法

+++ Competition Code/Point.h
@@ -8,7 +8,7 @@
 
 
 typedef double angle;
-const int PI = acos(-1);
+const double PI = acos(-1);
 
 
 // 结构体：平面上的点
